+++ src/modules/backgroundCheck_mod.f90
@@ -629,6 +629,8 @@ module backgroundCheck_mod
     real(8), parameter :: zzdcrit(3) = (/  9.00D0, 16.00D0, 25.00D0 /)
     real(8), parameter :: zchcrit(3) = (/  9.00D0, 16.00D0, 25.00D0 /)
     real(8), parameter :: zLogViscrit(3) = (/ 10.00D0, 20.00D0, 30.00D0 /)
+    ! Temporary hardcoded values for radar Doppler velocity
+    real(8), parameter :: radvelcrit(3) = (/ 8.00D0, 20.00D0, 30.00D0 /)
 
     !real(8) :: zuvcrit(3) = (/ 10.00D0, 20.00D0, 30.00D0 /)
     real(8) :: zuvcrit(3)
@@ -823,6 +825,17 @@ module backgroundCheck_mod
         isetflag =3
       end if
 
+    else if ( kvnam == bufr_radvel ) then
+    
+      ! Set flag for  Doppler Velocity (Radvel)
+      if ( zbgchk >= radvelcrit(1) .and. zbgchk < radvelcrit(2) ) then
+        isetflag = 1
+      else if ( zbgchk >= radvelcrit(2) .and. zbgchk < radvelcrit(3) ) then
+        isetflag = 2
+      else if ( zbgchk >= radvelcrit(3) ) then
+        isetflag = 3
+      end if
+
     else if ( obsFamily == 'CH' ) then
     
       ! SET FLAG FOR CHEMICAL CONSTITUENTS
@@ -833,8 +846,8 @@ module backgroundCheck_mod
       else if ( zbgchk >= zchcrit(3) ) then
         isetflag =3
       end if
-         
-    end if
+
+    end if 
 
     return